requirements.txt
fastapi==0.85.0
uvicorn==0.18.3
beanie==1.11.9
pymongo[srv]==4.2.0
python-dotenv
